+++ ios/Classes/BannerAdFactory.swift
@@ -1,0 +1,28 @@
+import Flutter
+import UnityAds
+
+class BannerAdFactory: NSObject, FlutterPlatformViewFactory {
+    private var messenger: FlutterBinaryMessenger
+    
+    init(messenger: FlutterBinaryMessenger) {
+        self.messenger = messenger
+        super.init()
+    }
+    
+    func createArgsCodec() -> FlutterMessageCodec & NSObjectProtocol {
+        return FlutterStandardMessageCodec.sharedInstance()
+    }
+    
+    func create(
+        withFrame frame: CGRect,
+        viewIdentifier viewId: Int64,
+        arguments args: Any?
+    ) -> FlutterPlatformView {
+        return BannerAdView(
+            frame: frame,
+            id: viewId,
+            arguments: args,
+            messenger: messenger)
+    }
+    
+}