+++ ios/Classes/UnityAdsConstants.swift
@@ -1,0 +1,28 @@
+struct UnityAdsConstants {
+    static let MAIN_CHANNEL = "unity.ads";
+    
+    static let BANNER_AD_CHANNEL = MAIN_CHANNEL + "/bannerAd";
+    static let VIDEO_AD_CHANNEL = MAIN_CHANNEL + "/videoAd";
+    
+    static let GAME_ID_PARAMETER = "gameId";
+    static let TEST_MODE_PARAMETER = "testMode";
+    static let FIREBASE_TEST_LAB_MODE_PARAMETER = "firebaseTestLabMode";
+    
+    static let PLACEMENT_ID_PARAMETER = "placementId";
+    static let HEIGHT_PARAMETER = "height";
+    static let WIDTH_PARAMETER = "width";
+    
+    static let INIT_METHOD = "init";
+    static let IS_READY_METHOD = "isReady";
+    static let SHOW_VIDEO_METHOD = "showVideo";
+    
+    static let READY_METHOD = "ready";
+    static let START_METHOD = "start";
+    static let COMPLETE_METHOD = "complete";
+    static let SKIPPED_METHOD = "skipped";
+    static let ERROR_METHOD = "error";
+    
+    static let BANNER_ERROR_METHOD = "banner_error";
+    static let BANNER_LOADED_METHOD = "banner_loaded";
+    static let BANNER_CLICKED_METHOD = "banner_clicked";
+}